MQTT: add basic send-only support
This lets you publish a message to a topic as a "power" command.

Currently, only a single power command is supported, and subscribing to topics is not yet supported.

For the future, subscribing would be useful to get confirmation from the device that it received the command.

For now, we return success to the user if the MQTT broker received the message, but this doesn't guarantee the device actually received it.