MQTT: add receive validation for expected values
After sending a command over MQTT, it's common for a device to publish a
status message to a topic which can be used to determine success.

Without this, you might "successfully" publish a message to MQTT, but
unless the device responds, you have no guarantee it worked.

This implementation waits for a matching topic + message to be
published. A matching message indicates success. Failure is indicated by
a timeout, which must be kept short (<5s) to remain user-visible.

Receive topics are subscribed while a request is in-flight, and are
unsubscribed when done.

The implementation would have been simpler by remaining subscribed to
any topics indefinitely, but that would have been inconsistent with the
spirit of overengineering so prevalent in this project.