MQTT: handle disconnects gracefully, simplify init
The readLoop() goroutine was only instantiated once, but returned
if an error occurred, never starting again.

Events as simple as the MQTT connection disconnecting (even if
reconnecting would otherwise work) cause an error here, so we
need to restart readLoop() if this happens.

We wait until the next MQTT operation before restarting the loop,
since it's only worthwhile doing so until we know a user is present,
and enough time has passed for the (transient) error to clear.