MQTT: unblock read loop, buffer received messages with cap
Most issues I’ve had with the MQTT implementation so far have been around deadlocks and blocking the read loop.

In particular, the go-mqtt implementation depends on ReadSlices() not just for receiving messages, but acts as a fundamental event loop where it internally reads and acks various control messages.

The best way to improve reliability here is to remove the ability for the read loop to block. As such, rather than delivering to subscribers unbuffered, it uses a channel to internally buffer messages, and if that buffer is exceeded, it simply drops the messages.

If you get into a full buffer situation, things have gone sufficiently wrong that it’s not useful to queue the messages any further.