MQTT: vastly simplify by removing unsubscribes
The implementation was overly complex, full of race conditions, and difficult to read and understand.

Most of the complexity is due to trying to handle unsubscribes, which is simply not a worthwhile optimisation for a small bandwidth reduction but large amount of complexity.

It's still a bit over the top, but at least it's simpler than it was.