Guest: add support for time-limited guest invites
A useful feature is the ability to have visitors let themselves in.

Without a dedicated guest invite feature, users may be tempted to share usernames/passwords with guests, which has the risk of allowing permanent access.

Instead, these guest invites are time-limited to a configurable duration. Effectively, a guest pass is a special type of session with a shorter expiry time than usual.

The guest pass is a URL, with the token passed as the ?t=<token> query parameter, which makes it copy-pastable to any messaging service your guest happens to use, and instantly logs them on.

Because the host may want to test the guest URL before sending, it takes precedence over any cookies already set. But, to avoid accidentally logging the host off, no cookies are touched by this, so reverting back to the non-token URL behaves like before.

While invalid sessions have previously been handled by simply showing the login form (after all, why show an error when you can jump straight to the expected action), there is some special treatment of invalid guest passes.

To avoid an expired guest pass resulting in a login screen, which would be confusing, a permanent error is displayed saying the pass is expired. Since there's nothing the guest can do at this point, there's no suggested action.

Guests have the same effective group memberships as their host. This is cached within the session entry, so the underlying auth mechanism doesn't need to be aware of the guest.

Because it's possible guests shouldn't have access to some things the host can access, an implicit "guest" group membership is added. ACLs can restrict guests by explicitly denying the user or group "guest".