Guest: return lifetime in Index, hide if unset
If the guest lifetime is unset, that implies guest invites are disabled.

However, we need to return this via the index so we can hide the guest invite button.

Since the invite lifetime isn't secret, we return the concrete value in the index response.

However, the frontend logic currently only checks whether or not the value is zero.