Index: add unit tests for handler
With the addition of the guest invites, which might/mightn't be present depending on configs, and devices that might be visible depending on ACLs, it's worth making sure the index is correct.

The index is the primary data view consumed by the user, with lots of inputs, so accidentally including things we shouldn't, or omitting something is a very real possibility.

Unit tests give us a chance to notice this before the user does.