Improve favicon: set explicit background
Various browsers render the icon differently. Most desktop browsers render with a white background, while Firefox renders with a black background when installed as a PWA.

Additionally, Firefox prefers the SVG for the PWA icon, but when installed on an Android home screen, it gets cropped too much.

I've adjusted it to be explicitly black foreground on white background, and increased the margins for the SVG.

I won't pretend to fully understand the root causes, nor can I guarantee this renders correctly in all situations.