Session: look for newest cookie, not just first
Sometimes, a user has stale cookies for one reason or another.

Unfortunately, http.Request.Cookie() only returns the first cookie found, which might not be valid.

This change deals with the problem defensively.

First, all cookies are examined, and the newest is picked. Chances are, the newest one corresponds to the most recent login.

Secondly, for OAuth, cookies are always explicitly expired if the user needs to log in again.